requirements.txt
fastapi
uvicorn
telethon
pymongo
python-dotenv
fastapi[all]